+++ apps/api/views/branch_views.py
@@ -1,9 +1,15 @@
 import logging
 import sys
 
+from logging import Logger
+
+from django.core.handlers.wsgi import WSGIRequest
+from django.db.models import QuerySet
+from django.http import QueryDict
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
+from rest_framework.serializers import Serializer
 from rest_framework import status
 
 from drf_yasg import openapi
@@ -16,8 +22,9 @@ from ..serializers import BranchSerializer
 
 from ..services.company_services import get_company_by_id_service
 from ..services.branch_service import get_branch_by_id_service, create_branch_by_company
+from ...core.models import Company, User, Branch
 
-logger = logging.getLogger('django')
+logger: Logger = logging.getLogger('django')
 
 
 @swagger_auto_schema(
@@ -57,19 +64,19 @@ logger = logging.getLogger('django')
 )
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
-def read_branch_list_view(request, pk):
+def read_branch_list_view(request: WSGIRequest, pk: int) -> Response:
     """Контроллер для отдачи информации о branch"""
-    user = request.user
+    user: QuerySet[User] = request.user
 
     try:
-        company = get_company_by_id_service(user, pk)
-        if company:
-            branch_list = company.branchs.all()
-            query = request.query_params.get('keyword')
+        company: QuerySet[Company] | None = get_company_by_id_service(user, pk)
+        if not(company is None):
+            branch_list: QuerySet[Branch] = company.branchs.all()
+            query: str = request.query_params.get('keyword')
             if query is None:
                 query = ''
-            page = request.query_params.get('page')
-            paginator = Paginator(branch_list, 10)
+            page: int = request.query_params.get('page')
+            paginator: Paginator = Paginator(branch_list, 10)
 
             try:
                 branch_list = paginator.page(page)
@@ -79,22 +86,22 @@ def read_branch_list_view(request, pk):
                 branch_list = paginator.page(paginator.num_pages)
 
             if page is None:
-                page = 1
+                page: int = 1
 
             page = int(page)
-            serializer = BranchSerializer(branch_list, many=True)
-            message = 'Запрос выполнен успешно'
+            serializer: Serializer[BranchSerializer] = BranchSerializer(branch_list, many=True)
+            message: str = 'Запрос выполнен успешно'
             logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(
                 data={
-                    'reviews': serializer.data,
+                    'branches': serializer.data,
                     'page': page,
                     'pages': paginator.num_pages
                 },
                 status=status.HTTP_200_OK
             )
         else:
-            message = 'Это не ваша company'
+            message: str = 'Это не ваша company'
             logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(
                 data={
@@ -104,7 +111,7 @@ def read_branch_list_view(request, pk):
             )
 
     except ObjectDoesNotExist as e:
-        message = f'Такой company не найдено {e}'
+        message: str = f'Такой company не найдено {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={
@@ -152,22 +159,22 @@ def read_branch_list_view(request, pk):
 )
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
-def read_branch_view(request, pk):
+def read_branch_view(request: WSGIRequest, pk: int) -> Response:
     """Контроллер для отдачи информации о branch"""
-    user = request.user
+    user: QuerySet[User] = request.user
 
     try:
-        branch = get_branch_by_id_service(user=user, branch_id=pk)
-        if branch:
-            serializer = BranchSerializer(branch, many=False)
-            message = 'Запрос выполнен успешно'
+        branch: QuerySet[Branch] | None = get_branch_by_id_service(user=user, branch_id=pk)
+        if not(branch is None):
+            serializer: Serializer[BranchSerializer] = BranchSerializer(branch, many=False)
+            message: str = 'Запрос выполнен успешно'
             logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(
                 data=serializer.data,
                 status=status.HTTP_200_OK
             )
         else:
-            message = 'Это не ваш branch'
+            message: str = 'Это не ваш branch'
             logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(
                 data={
@@ -177,7 +184,7 @@ def read_branch_view(request, pk):
             )
 
     except ObjectDoesNotExist as e:
-        message = f'Такой branch не найден {e}'
+        message: str = f'Такой branch не найден {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={
@@ -187,7 +194,7 @@ def read_branch_view(request, pk):
         )
 
     except Exception as e:
-        message = f'Ошибка при обработке запроса {e}'
+        message: str = f'Ошибка при обработке запроса {e}'
         logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={
@@ -280,26 +287,27 @@ def read_branch_view(request, pk):
 )
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
-def create_branch_view(request):
+def create_branch_view(request: WSGIRequest) -> Response:
     """Контроллер для создания review"""
-    user = request.user
+    user: QuerySet[User] = request.user
+    data: QueryDict = request.data
 
     try:
-        company_id = request.data['company_id']
-        name = request.data['name']
-        branch = create_branch_by_company(user=user, name=name, company_id=company_id)
-        if branch:
-            serializer = BranchSerializer(branch, data=request.data, partial=True)
+        company_id: int = data['company_id']
+        name: str = data['name']
+        branch: QuerySet[Branch] | None = create_branch_by_company(user=user, name=name, company_id=company_id)
+        if not(branch is None):
+            serializer: Serializer[BranchSerializer] = BranchSerializer(branch, data=request.data, partial=True)
             serializer.is_valid()
             serializer.save()
-            message = 'Запрос выполнен успешно'
+            message: str = 'Запрос выполнен успешно'
             logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(
                 serializer.data,
                 status=status.HTTP_201_CREATED
             )
         else:
-            message = 'Это не ваш company'
+            message: str = 'Это не ваш company'
             logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(
                 data={
@@ -309,7 +317,7 @@ def create_branch_view(request):
             )
 
     except KeyError as e:
-        message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
+        message: str = f'Ошибка при обработке запроса. Отсутствует поле {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={
@@ -319,7 +327,7 @@ def create_branch_view(request):
         )
 
     except ObjectDoesNotExist as e:
-        message = f'Такой company не найден {e}'
+        message: str = f'Такой company не найден {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={
@@ -329,7 +337,7 @@ def create_branch_view(request):
         )
 
     except Exception as e:
-        message = f'Ошибка при обработке запроса {e}'
+        message: str = f'Ошибка при обработке запроса {e}'
         logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={
@@ -415,25 +423,25 @@ def create_branch_view(request):
 )
 @api_view(['PUT'])
 @permission_classes([IsAuthenticated])
-def update_branch_view(request, pk):
+def update_branch_view(request: WSGIRequest, pk: int) -> Response:
     """Контроллер для обновления информации branch"""
-    user = request.user
-    data = request.data
+    user: QuerySet[User] = request.user
+    data: QueryDict = request.data
 
     try:
-        branch = get_branch_by_id_service(user, pk)
-        if branch:
-            serializer = BranchSerializer(branch, many=False, partial=True, data=data)
+        branch: QuerySet[Branch] | None = get_branch_by_id_service(user, pk)
+        if not(branch is None):
+            serializer: Serializer[BranchSerializer] = BranchSerializer(branch, many=False, partial=True, data=data)
             if serializer.is_valid():
                 serializer.save()
-            message = 'Запрос выполнен успешно'
+            message: str = 'Запрос выполнен успешно'
             logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(
                 data=serializer.data,
                 status=status.HTTP_200_OK
             )
         else:
-            message = 'Это не ваш branch'
+            message: str = 'Это не ваш branch'
             logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(
                 data={
@@ -443,7 +451,7 @@ def update_branch_view(request, pk):
             )
 
     except ObjectDoesNotExist as e:
-        message = f'Такой branch не найден {e}'
+        message: str = f'Такой branch не найден {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={
@@ -453,7 +461,7 @@ def update_branch_view(request, pk):
         )
 
     except KeyError as e:
-        message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
+        message: str = f'Ошибка при обработке запроса. Отсутствует поле {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={
@@ -463,7 +471,7 @@ def update_branch_view(request, pk):
         )
 
     except Exception as e:
-        message = f'Ошибка при обработке запроса {e}'
+        message: str = f'Ошибка при обработке запроса {e}'
         logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={