+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added django-silk [@fringer2423](https://github.com/fringer2423)
 - Configured silk
 - Typed answer views
+- Typed branch views
 
 ### Изменено